Fix for the issue https://github.com/epam/ketcher/issues/3951
When a molecule name is null we now don't print out 'null' but print an empty string instead.

This is in accordance with the specification:
A compound record contains several distinct sections. First, there is a three-line header block. These three lines may contain:

The name of the molecule
Details of the software used to generate the compound structure
A comment
Alternatively, any (or all) of these lines may be left blank.